Refactor Comments models
