+++ src/domain/shared/components/Comments/CommentsComponent.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { FetchResult } from '@apollo/client';
 import { Box, Typography } from '@mui/material';
 import DashboardGenericSection from '../DashboardSections/DashboardGenericSection';
-import { Comment } from '../../../../models/discussion/comment';
+import { Comment } from './models/comment';
 import { MID_TEXT_LENGTH } from '../../../../models/constants/field-length.constants';
 import { mapWithSeparator } from '../../utils/joinNodes';
 import SectionSpacer from '../Section/SectionSpacer';
