+++ src/domain/shared/components/Comments/MessageView.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { DeleteOutlined } from '@mui/icons-material';
 import { Box, Typography, styled, BoxProps, IconButton, Grid } from '@mui/material';
-import { Comment } from '../../../../models/discussion/comment';
+import { Comment } from './models/comment';
 import Markdown from '../../../../common/components/core/Markdown';
 import { formatCommentDate } from './formatCommentDate';
 import AuthorAvatar from './AuthorAvatar';
