+++ src/views/Discussions/DiscussionView.tsx
@@ -6,8 +6,8 @@ import Filter from '../../domain/admin/components/Common/Filter';
 import MessageView from '../../domain/shared/components/Comments/MessageView';
 import PostMessageToCommentsForm from '../../domain/shared/components/Comments/PostMessageToCommentsForm';
 import { MID_TEXT_LENGTH } from '../../models/constants/field-length.constants';
-import { Comment } from '../../models/discussion/comment';
-import { Discussion } from '../../models/discussion/discussion';
+import { Comment } from '../../domain/shared/components/Comments/models/comment';
+import { Discussion } from '../../domain/discussion/models/discussion';
 import { AuthorizationPrivilege } from '../../models/graphql-schema';
 
 export interface DiscussionViewProps {
